almost final state pending:top up credit registration_system_responses need fixes

+++ header_file/socket_definitions.h
@@ -32,7 +32,7 @@ typedef struct data {
     char userName[USER_LENGTH + 1];
     pthread_mutex_t mutex;
     int socket;
-    int state;
+    size_t state;
     int stop;
 } DATA;
